modif dijkstra

+++ be-graphes-algos/src/main/java/org/insa/graphs/algorithm/shortestpath/DijkstraAlgorithm.java
@@ -101,7 +101,7 @@ public class DijkstraAlgorithm extends ShortestPathAlgorithm {
                 if (label_destinataire.getMarque() == false) { // si le sommet n'est pas marqué
                 // affecter comme coût à ce sommet le minimum entr son coût et le poids de l'arc
                     old_cost = label_destinataire.getCost() ;
-                    new_cost = min_tas.getCost() + (int)data.getCost(arc) ;
+                    new_cost = min_tas.getCost() + data.getCost(arc) ;
 
                     /*if (Double.isInfinite(old_cost) && Double.isFinite(new_cost)) {
                         notifyNodeReached(arc.getDestination());
